Added engine simulator to GIT
Engine simulator calculates engine speed and torque from throttle position and load. It outputs a missing tooth pulse waveform corresponding to the engine speed.

It receives a fuel input pulse and measures the length. It uses this to calculate the corresponding Air to Fuel Ratio. The AFR is used to calculate a lambda value, which is then output on an analog output pin for the ECU to read.

The AFR affects the maximum torque output of the engine